set current spec
